Workaround assert failure in sys_framesync
This assertion was never reasonable (scheduler nonsense could trip
this..), nevertheless I'd never seen it fail until the last few weeks
when it failed twice. I haven't given any thought as to why this would
occur - commenting it out until I can investigate. On both occasions
so far it triggered about five minutes into a game of FF1, and shortly
after waking the laptop from a long rest. Changing it to a warning.