Quick hacks tweaking timer layout for my twitch stream
I want to move/resize the timer so it doesn't overlap the NES output
(so that the emulator can be cropped out for restreaming without bits
of the timer showing), and eventually separate the armed/paused text
from the counter itself. This commit removes the armed/paused text
so the timer state is only distinguished by color, and that in the
most quick and dirty way.

I'm not sure what the right way to deal with this possible is because
the window size needs to be increased to make room for the timer..