Tiny reduction of keyboard input latency
When the emulated machine strobes the controller shift
register to read input, immediately pump the event loop
and peek for keyboard events, processing them if they
correspond to controller input keys. This eliminates up
to a frame of input lag in some cases, as games will
frequently poll for input after vblank is processed,
at the very end of the emulated frame.

In practice, this case is not hit that often, because
the emulator is usually running much faster than realtime
and sleeping at the end of vblank.

This probably won't save me more than a second or two per
run, but I had to implement in order to measure the benefit..

I should do something similar for joypad inputs, but I'm guessing
it's not worth the effort.