Varying degrees of nitpicky stuff to fix build on Apple Silicon
Fix some warnings and attempt #ifdef fallback to the plain C paths for
the NTSC output's FIR filters. Since I haven't attempted subsequently
rebuilding on Intel, I may have broken that build.

Extend M6502 cycle counting to use a long long so that, hopefully, we
never encounter a wraparound under reasonable conditions. This is the
lazy path, vs. changing it to unsigned arithmetic and having to think
harder (it is definitely being used in its twos-complement signed
integer capacity). Should be good enough. Something was badly
breaking the emulator initially. I think, though I'm not 100% sure,
that it's due to the cycles counter wrapping around, and recent clang
doing some UB optimization bullshit. This is the major thing that
popped out under -fsanitize=undefined, and what seemed to fix the
problems. Reassuring, since I first anticipated this could be
scattered across the 6502 emulator, I'd be down to comparing execution
traces to fix the bugs. Good thing it didn't come to that.